improvement: add Preview button to AI apply confirmation dialog
- Extract WorkflowOverview from OverviewMode for reuse
- Add Preview button to DiffPreviewDialog footer
- Show proposed workflow in z-index 10002 overlay
- Document new DIALOG_PREVIEW_OVERLAY tier in CLAUDE.md
