+++ src/webview/src/i18n/translations/ko.ts
@@ -415,6 +415,8 @@ export const koWebviewTranslations: WebviewTranslationKeys = {
     '경고: AI가 워크플로우를 가져온 후 캔버스가 수정되었습니다. 변경 사항을 신중하게 검토한 후 수락하세요.',
   'dialog.diffPreview.applyAnyway': '변경 사항 적용',
   'dialog.diffPreview.retryWithLatest': '최신으로 재시도',
+  'dialog.diffPreview.previewOverview': '미리보기',
+  'dialog.diffPreview.closeOverview': '미리보기 닫기',
 
   // Reset Workflow Confirmation Dialog
   'toolbar.resetWorkflow': '워크플로우 초기화',
