+++ src/webview/src/i18n/translations/zh-TW.ts
@@ -403,6 +403,8 @@ export const zhTWWebviewTranslations: WebviewTranslationKeys = {
     '警告：AI取得工作流程後畫布已被修改。請仔細檢查變更內容後再接受。',
   'dialog.diffPreview.applyAnyway': '仍然套用',
   'dialog.diffPreview.retryWithLatest': '使用最新重試',
+  'dialog.diffPreview.previewOverview': '預覽',
+  'dialog.diffPreview.closeOverview': '關閉預覽',
 
   // Reset Workflow Confirmation Dialog
   'toolbar.resetWorkflow': '重設工作流程',
